test(ui): split swarm activity integration coverage
