+++ ui/src/lib/sessions/index.test.ts
@@ -5,7 +5,7 @@ import {
   type GatewayEventFrame,
   type GatewayHelloOk,
 } from "../../api/gateway.ts";
-import type { GatewaySessionRow, SessionsListResult } from "../../api/types.ts";
+import type { SessionsListResult } from "../../api/types.ts";
 import { waitForFast } from "../../test-helpers/wait-for.ts";
 import { createSessionCapability, reconcileSessionRunTerminal } from "./index.ts";
 
@@ -1077,127 +1077,6 @@ describe("createSessionCapability", () => {
       }),
     );
     expect(request).toHaveBeenCalledTimes(2);
-    sessions.dispose();
-  });
-
-  it("keeps chronological swarm phase and log annotations across canonical refreshes", async () => {
-    const parentKey = "agent:main:main";
-    const groupId = "swarm:agent:main:main:turn-42";
-    let rows: SessionsListResult["sessions"] = [
-      { key: parentKey, kind: "direct", updatedAt: 1 },
-      {
-        key: "agent:main:subagent:older",
-        kind: "direct",
-        parentSessionKey: parentKey,
-        swarmGroupId: groupId,
-        status: "running",
-        updatedAt: 2,
-      },
-    ];
-    let ts = 1;
-    const request = vi.fn(async (method: string) => {
-      if (method !== "sessions.list") {
-        throw new Error(`Unexpected request: ${method}`);
-      }
-      return sessionsResult(rows, ts++);
-    });
-    const client = { request } as unknown as GatewayBrowserClient;
-    const { gateway, emitEvent } = createGatewayHarness(client);
-    const sessions = createSessionCapability(gateway);
-    const note = (kind: "phase" | "log", text: string) => ({
-      sessionKey: parentKey,
-      reason: "swarm-note",
-      swarmGroupId: groupId,
-      kind,
-      text,
-      key: parentKey,
-      updatedAt: 1,
-    });
-    const child = (key: string, updatedAt: number) => ({
-      sessionKey: key,
-      reason: "create",
-      key,
-      kind: "direct",
-      parentSessionKey: parentKey,
-      swarmGroupId: groupId,
-      status: "running",
-      updatedAt,
-    });
-    const emitChanged = (payload: Record<string, unknown>) =>
-      emitEvent({ type: "event", event: "sessions.changed", payload });
-
-    await sessions.refresh({ force: true });
-    emitChanged(note("phase", "Plan"));
-    await waitForFast(() => expect(request).toHaveBeenCalledTimes(2));
-    rows = [
-      ...rows,
-      {
-        key: "agent:main:subagent:planner",
-        kind: "direct",
-        parentSessionKey: parentKey,
-        swarmGroupId: groupId,
-        status: "running",
-        updatedAt: 3,
-      },
-    ];
-    emitChanged(child("agent:main:subagent:planner", 3));
-    await waitForFast(() =>
-      expect(sessions.state.result?.sessions.some((row) => row.key.endsWith(":planner"))).toBe(
-        true,
-      ),
-    );
-
-    type SwarmDisplayRow = GatewaySessionRow & { swarmLog?: string; swarmPhase?: string };
-    const displayRows = () => sessions.state.result?.sessions as SwarmDisplayRow[] | undefined;
-    expect(displayRows()?.find((row) => row.key.endsWith(":older"))?.swarmPhase).toBeUndefined();
-    expect(displayRows()?.find((row) => row.key.endsWith(":planner"))?.swarmPhase).toBe("Plan");
-
-    emitChanged(note("log", "Planning is complete."));
-    await waitForFast(() =>
-      expect(
-        displayRows()
-          ?.filter((row) => row.swarmGroupId === groupId)
-          .map((row) => row.swarmLog),
-      ).toEqual(["Planning is complete.", "Planning is complete."]),
-    );
-
-    emitChanged(note("phase", "Build"));
-    rows = [
-      ...rows,
-      {
-        key: "agent:main:subagent:builder",
-        kind: "direct",
-        parentSessionKey: parentKey,
-        swarmGroupId: groupId,
-        status: "running",
-        updatedAt: 4,
-      },
-    ];
-    emitChanged(child("agent:main:subagent:builder", 4));
-    await waitForFast(() =>
-      expect(sessions.state.result?.sessions.some((row) => row.key.endsWith(":builder"))).toBe(
-        true,
-      ),
-    );
-
-    expect(displayRows()?.find((row) => row.key.endsWith(":planner"))?.swarmPhase).toBe("Plan");
-    expect(displayRows()?.find((row) => row.key.endsWith(":builder"))?.swarmPhase).toBe("Build");
-
-    // A pre-phase child receiving a plain status update after phases were
-    // announced must stay Unphased; only creation events assign implicitly.
-    emitChanged({
-      sessionKey: "agent:main:subagent:older",
-      reason: "status",
-      key: "agent:main:subagent:older",
-      kind: "direct",
-      parentSessionKey: parentKey,
-      swarmGroupId: groupId,
-      status: "done",
-      updatedAt: 5,
-    });
-    await waitForFast(() =>
-      expect(displayRows()?.find((row) => row.key.endsWith(":older"))?.swarmPhase).toBeUndefined(),
-    );
     sessions.dispose();
   });
 });
